(NSW) We now have only one non-threaded audio driver for both SDKs instead of two
